requirements.txt
monai
pytorch-lightning
torch
tensorboard
nibabel
pyyaml